+++ gulpfile.js/index.js
@@ -1,0 +1,14 @@
+/**
+ *
+ * Gulpfile
+ *
+ * Copyright 2016, Web Framework by Stewart Dellow
+ * MIT License (MIT)
+ *
+**/
+
+var requireDir = require('require-dir');
+
+requireDir('./tasks', {
+    recurse: true
+});