+++ gulpfile.js/tasks/sample.js
@@ -1,0 +1,48 @@
+/**
+ *
+ * Gulpfile > Tasks > Sample
+ *
+ * Copyright 2016, Web Framework by Stewart Dellow
+ * MIT License (MIT)
+ *
+**/
+
+;(function(GulpTask){
+
+	/* =========================================================================== */
+	/* Config
+	/* =========================================================================== */
+	// var config = require('../config.json');
+
+
+	/* =========================================================================== */
+	/* Global Dependencies
+	/* =========================================================================== */
+	// var gulp   = require('gulp'),
+	// 	args   = require('yargs').argv,
+	// 	notify = require('gulp-notify'),
+	// 	buffer = require('vinyl-buffer'),
+	// 	source = require('vinyl-source-stream');
+
+
+	/* =========================================================================== */
+	/* Task Dependencies
+	/* =========================================================================== */
+	// 	var module = require('gulp-module');
+
+
+	/* =========================================================================== */
+	/* Task
+	/* =========================================================================== */
+	// GulpTask = function(){
+	//     return gulp.src(config.src)
+	//         .pipe(module());
+	// }
+	//
+	// // Task
+	// gulp.task('sample', GulpTask);
+	//
+	// // Export
+	// module.exports = GulpTask;
+
+}(function(){}));